Removed spectrum elements from the transport state
