+++ tardis/spectrum/spectrum.py
@@ -1,6 +1,8 @@
 import warnings
+
 import numpy as np
 from astropy import units as u
+
 from tardis.io.util import HDFWriterMixin
 
 
@@ -37,9 +39,9 @@ class TARDISSpectrum(HDFWriterMixin):
         self._frequency = _frequency.to("Hz", u.spectral())
         self.luminosity = luminosity.to("erg / s")
 
-        l_nu_unit = u.def_unit("erg\ s^{-1}\ Hz^{-1}", u.Unit("erg/(s Hz)"))
+        l_nu_unit = u.def_unit(r"erg s^{-1} Hz^{-1}", u.Unit("erg/(s Hz)"))
         l_lambda_unit = u.def_unit(
-            "erg\ s^{-1}\ \\AA^{-1}", u.Unit("erg/(s AA)")
+            r"erg s^{-1} \AA^{-1}", u.Unit("erg/(s AA)")
         )
 
         self.frequency = self._frequency[:-1]
